+++ changelog.rst
@@ -5,6 +5,8 @@ Change Log
 =========================
 
 * add include_rst() and include_mkdn() as functions available in Jinja templates
+* css and js from included templates is now always included, regardless of where
+    in the child template the include method is called.
 
 0.4.4 released 2011-11-09
 =========================