+++ changelog.rst
@@ -7,6 +7,8 @@ Change Log
 * add include_rst() and include_mkdn() as functions available in Jinja templates
 * css and js from included templates is now always included, regardless of where
     in the child template the include method is called.
+* add "content" template filter which will merge css/jss from the given content
+  object into the parent template
 
 0.4.4 released 2011-11-09
 =========================